Refactor skip list
