+++ pysearchlite/test_spim_inverted_index_skip_list_memory.py
@@ -1,7 +1,6 @@
 import io
 import os.path
 import tempfile
-from random import randrange
 
 import pytest
 
@@ -10,9 +9,6 @@ from .spim_inveted_index_skip_list_memory import (
     read_token,
     write_token,
     write_doc_ids,
-    make_skip_list,
-    skip_list_search,
-    skip_list_and,
 )
 
 
@@ -149,54 +145,3 @@ def test_spim_count_and(spim_index):
 def test_spim_clear(spim_index):
     assert spim_index.raw_data == {}
     assert spim_index.data == {}
-
-
-def linear_search(arr, target, left=0, right=None):
-    if right is None:
-        right = len(arr)
-    for i in range(left, right):
-        if arr[i] >= target:
-            return i
-    return right
-
-
-def skip_list_search_test_cases(num, max_len):
-    tests = []
-    for _ in range(num):
-        arr_len = randrange(1, max_len)
-        arr = sorted(set(randrange(arr_len * 4) for _ in range(arr_len)))
-        target = randrange(arr_len * 4)
-        tests.append((arr, target))
-    return tests
-
-
-@pytest.mark.parametrize('arr, target', skip_list_search_test_cases(100, 10))
-def test_skip_list_search(arr, target):
-    s = make_skip_list(arr)
-    ret, skip_list_pos = skip_list_search(s,target, 0)
-    i = linear_search(arr, target)
-    if i == len(arr):
-        assert ret == arr[-1]
-    else:
-        assert ret == arr[i]
-
-
-def skip_list_and_test_cases(num, max_len):
-    tests = []
-    for _ in range(num):
-        a_len = randrange(1, max_len)
-        b_len = randrange(1, max_len)
-        if a_len > b_len:
-            a_len, b_len = b_len, a_len
-        a = sorted(set(randrange(a_len * 4) for _ in range(a_len)))
-        b = sorted(set(randrange(b_len * 4) for _ in range(b_len)))
-        tests.append((a, b))
-    return tests
-
-
-@pytest.mark.parametrize('a, b', skip_list_and_test_cases(100, 10))
-def test_skip_list_and(a, b):
-    s = make_skip_list(a)
-    t = make_skip_list(b)
-    result = skip_list_and(s, t)
-    assert set(result) == set(a) & set(b)
